requirements.txt
pytest
setuptools
wheel
setuptools_scm[toml]
boto3
matplotlib
pandas
streamlit
plotly
azure-core
azure-ai-textanalytics
azure-identity
